src/sample_flow/requirements.txt
openai
azure-identity
promptflow[azure]==1.11.0
promptflow-tracing==1.11.0
promptflow-tools==1.11.0
python-dotenv